[refactor] 파일명 camelCase로 통일
