+++ src/main/base/common/fileUtils.ts
