+++ src/main/base/common/pathUtils.ts
